requirements.txt
google-adk
google-generativeai
rouge-score
sacrebleu
